+++ cypress/integration/components/checkbox.js
@@ -1,0 +1,7 @@
+const name = 'checkbox';
+
+describe("Metro 4 :: Checkbox", () => {
+    it('Component Initialization', ()=>{
+        cy.visit("cypress/"+name+".html");
+    })
+})