+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 + [x] Toast: added method `Metro.createToast()`. This eq to `Metro.toast.create()`
 + [x] Components: event `on[Component]Create` now have context a `HTMLElement` and receive `$(HTMLElement)` as argument
 + [x] Keypad: fix Keypad backspace button click behavior on `keypad`, created at `runtime`, issue #1547
++ [x] Utils: improve function `isType`
++ [x] Clock: add events `onTick`, `onSecond`
++ [x] String: add extensions `includes()`, `camelCase()`, `dashedName()`, `shuffle()`
 
 
 ### 4.3.7